refactor(ownership): move ignored authority
