+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Client/Components/GroupComponent.Ownership.cs
@@ -5,9 +5,7 @@ namespace ShipCoreFramework
         private long GetObservedOwnerId()
         {
             if (_runtimeStateReceived) return _runtimeOwnerId;
-
-            var ownerId = ResolveLocalOwnerId();
-            return IsIgnoredNpcGroup() ? 0 : ownerId;
+            return ResolveLocalOwnerId();
         }
     }
 }
