+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Shared/Components/GroupComponent.Ownership.cs
@@ -1,21 +1,9 @@
 using System.Linq;
-using Sandbox.ModAPI;
-using VRage.Game;
-using VRage.Game.ModAPI;
 
 namespace ShipCoreFramework
 {
     internal partial class GroupComponent
     {
-        private bool IsIgnoredNpcGroup()
-        {
-            if (!Session.Config.IgnoreAiFactions) return false;
-
-            var mainGrid = MainCoreComponent?.CoreBlock?.CubeGrid;
-            return mainGrid?.IsNpcSpawnedGrid ??
-                   GridDictionary.Keys.Any(grid => grid != null && grid.IsNpcSpawnedGrid);
-        }
-
         internal long OwnerId
         {
             get
@@ -45,32 +33,8 @@ namespace ShipCoreFramework
 
         internal bool IsIgnoredGroup()
         {
-            if (!Session.IsServer && _runtimeStateReceived) return GetCachedIsIgnoredGroup();
+            if (!Session.IsServer) return GetCachedIsIgnoredGroup();
             return Session.IsGameThread ? ComputeIsIgnoredGroup() : GetCachedIsIgnoredGroup();
-        }
-
-        private bool ComputeIsIgnoredGroup()
-        {
-            if (IsIgnoredByAiOrFactionTag()) return true;
-            if (OwnerId == 0) return true;
-            var player = MyAPIGateway.Players.TryGetIdentityId(OwnerId);
-            return player != null && player.PromoteLevel == MyPromoteLevel.Admin &&
-                   MyAPIGateway.Session.IsUserIgnorePCULimit(player.SteamUserId);
-        }
-
-        internal bool IsIgnoredByAiOrFactionTag()
-        {
-            if (IsIgnoredNpcGroup()) return true;
-
-            var faction = OwningFaction;
-            if (faction == null) return false;
-            return Session.Config.IgnoredFactionTags != null &&
-                   Session.Config.IgnoredFactionTags.Contains(faction.Tag);
-        }
-
-        internal bool IsIgnoredByAiOrFactionTagThreadSafe()
-        {
-            return Session.IsGameThread ? IsIgnoredByAiOrFactionTag() : GetCachedIsIgnoredByAiOrFactionTag();
         }
 
         private long GetRepresentativeGridId()
